refactor: room refactor
