+++ client/src/components/Room/RoomInfo.tsx
@@ -2,8 +2,11 @@ import ExitButton from '../buttons/ExitButton';
 import { FaRegCopy } from 'react-icons/fa6';
 import { ToastContainer, toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
+import { useRoom } from '../../hooks/useRoom';
 
-export default function RoomInfo({ roomCode }: { roomCode: string }) {
+export default function RoomInfo() {
+  const { roomCode } = useRoom();
+
   const copyToClipboard = () => {
     navigator.clipboard.writeText(roomCode);
     toast.success('Copied to clipboard!', {
